Give the dense, compile-time-bounded state the type it actually has
A type audit of the contracts, following the tokenAddress fix. Same shape of
mistake in three more places, none of them dangerous, all of them costing gas or
a cast for nothing.

filledSubtrees and poolZeros were mappings keyed by level. The level space is
dense and LEVELS is a compile-time constant, so the index space IS the array — a
mapping pays a keccak to reach a slot the compiler can compute by addition, on
every level of every insert. SMTRegistry._smtZeros was already a fixed array,
which is what makes this an inconsistency rather than a decision.

_smtNodes was a mapping of mappings. Only the inner dimension is sparse; the
level is bounded at compile time, so an array of mappings hashes once per node
access instead of twice. Registration walks 32 levels reading two nodes each, so
that is 64 keccaks per registration.

AliasRecord.registeredAt was a uint64 that packs against nothing — it follows four
bytes32 fields, so it starts a fresh slot either way and bought only a
uint64(block.timestamp) cast at the single write. registryRootSeenAt already holds
a timestamp as uint256.

getSmtSiblings took a uint256 slot when aliasSlot is a uint32.

Measured, not assumed: transact 1,183,751 -> 1,181,971 and register 3,094,727 ->
3,091,037, via scripts/gasbench.ts, which is added here so the next such change is
judged rather than guessed.

Nothing changed in the rolling tree's own index. knownPoolRootTree is keyed by
root — a 32-byte hash, genuinely sparse — so a mapping is the right shape, and
filledSubtrees stays deliberately unkeyed by tree, which is what avoids paying ~16
zero-to-nonzero SSTOREs on whichever transaction opens a new tree. There are now
no nested mappings anywhere in the contracts.

The casts that remain are the ones that should: address(uint160(uint256(keccak)))
in the CREATE2 factories, where truncation is the definition of the address, and
the widening of tokenAddress back to a field element for the verifier's public
signals, which is total.

+++ packages/protocol/contracts/HaliasRegistry.sol
@@ -49,7 +49,11 @@ contract HaliasRegistry is SMTRegistry {
         bytes32 nullifierKeyHash;   // Poseidon(nullifierKey, 1) — the raw key is never on chain
         bytes32 encryptionPubkey;
         bytes32 dataHash;
-        uint64  registeredAt;
+        // uint256, not a narrower timestamp type. It follows four bytes32 fields, so it
+        // starts a fresh slot with nothing to pack against — the narrower type buys no
+        // storage and costs a `uint64(block.timestamp)` cast at the only write. Matches
+        // {SMTRegistry-registryRootSeenAt}, which is a timestamp for the same reason.
+        uint256 registeredAt;
     }
 
     mapping(bytes32 => AliasRecord) public aliases;
@@ -122,7 +126,7 @@ contract HaliasRegistry is SMTRegistry {
             nullifierKeyHash: nullifierKeyHash,
             encryptionPubkey: encryptionPubkey,
             dataHash:         bytes32(0),
-            registeredAt:     uint64(block.timestamp)
+            registeredAt:     block.timestamp
         });
 
         bytes32 leaf = _writeLeaf(aliasHash);
